ADR-2338 Scala 3 migration (#74)
* ADR-2338 Scala 3 migration

* ADR-2338 Improve coverage

+++ app/uk/gov/hmrc/alcoholdutyreturns/config/AppConfig.scala
@@ -42,7 +42,7 @@ class AppConfig @Inject() (config: Configuration, servicesConfig: ServicesConfig
   lazy val returnsGetReturnUrlFormat    = new MessageFormat(getConfStringAndThrowIfNotFound("returns.url.getReturn"))
   lazy val returnsSubmitReturnUrlFormat = new MessageFormat(getConfStringAndThrowIfNotFound("returns.url.submitReturn"))
 
-  val dbTimeToLiveInSeconds: Int    = config.get[Int]("mongodb.timeToLiveInSeconds")
+  val dbTimeToLiveInSeconds: Long   = config.get[Int]("mongodb.timeToLiveInSeconds")
   val lockingDurationInSeconds: Int = config.get[Int]("mongodb.lockingDurationInSeconds")
 
   def getSubscriptionSummaryUrl(appaId: String): String =
